Update the environment script to use environment variables if present
The build server won't have AWS credentials to retrieve the config from
SSM Parameter Store; we'll put the values in environment variables on
the build server. Now the update_environment_file script looks first for
existing environment variables and then for SSM Parameters if the
environment variables are missing. This supports local dev and the build
server.